OP works and is vastly cleaner inside.  The api has been modernized, but is not yet compatible with the old way.
